documentation/01-PROJECTS/cloudron-rd/Cloudron-R-et-D/AsyncDeepWiki/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
gitpython==3.1.40
beautifulsoup4==4.12.2
markdown==3.5.1
python-multipart==0.0.6
aiofiles==23.2.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==1.3.0
google-generativeai==0.3.0
anthropic==0.7.0
httpx==0.25.0
pypdf==3.17.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
langchain==0.0.354
langchain-community==0.0.13
chromadb==0.4.18
llama-index==0.9.15
adalflow==0.2.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
